+++ rbac-distributed-service-impl/src/main/java/com/dwarfeng/rbacds/impl/cache/PermissionMetaCacheImpl.java
@@ -1,0 +1,97 @@
+package com.dwarfeng.rbacds.impl.cache;
+
+import com.dwarfeng.rbacds.sdk.bean.entity.FastJsonPermissionMeta;
+import com.dwarfeng.rbacds.stack.bean.entity.PermissionMeta;
+import com.dwarfeng.rbacds.stack.bean.key.PermissionMetaKey;
+import com.dwarfeng.rbacds.stack.cache.PermissionMetaCache;
+import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
+import com.dwarfeng.subgrade.stack.exception.CacheException;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Repository
+public class PermissionMetaCacheImpl implements PermissionMetaCache {
+
+    private final RedisBatchBaseCache<PermissionMetaKey, PermissionMeta, FastJsonPermissionMeta> delegate;
+
+    public PermissionMetaCacheImpl(
+            RedisBatchBaseCache<PermissionMetaKey, PermissionMeta, FastJsonPermissionMeta> delegate
+    ) {
+        this.delegate = delegate;
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean exists(PermissionMetaKey key) throws CacheException {
+        return delegate.exists(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public PermissionMeta get(PermissionMetaKey key) throws CacheException {
+        return delegate.get(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void push(PermissionMeta value, long timeout) throws CacheException {
+        delegate.push(value, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void delete(PermissionMetaKey key) throws CacheException {
+        delegate.delete(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void clear() throws CacheException {
+        delegate.clear();
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(@SkipRecord List<PermissionMetaKey> keys) throws CacheException {
+        return delegate.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(@SkipRecord List<PermissionMetaKey> keys) throws CacheException {
+        return delegate.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    @SkipRecord
+    public List<PermissionMeta> batchGet(@SkipRecord List<PermissionMetaKey> keys) throws CacheException {
+        return delegate.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchPush(@SkipRecord List<PermissionMeta> entities, long timeout) throws CacheException {
+        delegate.batchPush(entities, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(@SkipRecord List<PermissionMetaKey> keys) throws CacheException {
+        delegate.batchDelete(keys);
+    }
+}