+++ rbac-distributed-service-node/src/main/java/com/dwarfeng/rbacds/node/configuration/FastJsonConfiguration.java
@@ -19,6 +19,7 @@ public class FastJsonConfiguration {
         ParserConfig.getGlobalInstance().addAccept(FastJsonPermission.class.getCanonicalName());
         ParserConfig.getGlobalInstance().addAccept(FastJsonUser.class.getCanonicalName());
         ParserConfig.getGlobalInstance().addAccept(FastJsonPermissionGroup.class.getCanonicalName());
+        ParserConfig.getGlobalInstance().addAccept(FastJsonPermissionMeta.class.getCanonicalName());
         LOGGER.debug("FastJson autotype 白名单配置完毕");
     }
 }