+++ rbac-distributed-service-sdk/src/main/java/com/dwarfeng/rbacds/sdk/bean/BeanMapper.java
@@ -1,7 +1,10 @@
 package com.dwarfeng.rbacds.sdk.bean;
 
 import com.dwarfeng.rbacds.sdk.bean.entity.*;
+import com.dwarfeng.rbacds.sdk.bean.key.FastJsonPermissionMetaKey;
+import com.dwarfeng.rbacds.sdk.bean.key.WebInputPermissionMetaKey;
 import com.dwarfeng.rbacds.stack.bean.entity.*;
+import com.dwarfeng.rbacds.stack.bean.key.PermissionMetaKey;
 import com.dwarfeng.subgrade.sdk.bean.key.FastJsonLongIdKey;
 import com.dwarfeng.subgrade.sdk.bean.key.FastJsonStringIdKey;
 import com.dwarfeng.subgrade.sdk.bean.key.JSFixedFastJsonLongIdKey;
@@ -43,6 +46,17 @@ public interface BeanMapper {
 
     @InheritInverseConfiguration
     LongIdKey longIdKeyFromWebInput(WebInputLongIdKey webInputLongIdKey);
+
+    // -----------------------------------------------------------Rbac Key-----------------------------------------------------------
+    FastJsonPermissionMetaKey permissionMetaKeyToFastJson(PermissionMetaKey permissionMetaKey);
+
+    @InheritInverseConfiguration
+    PermissionMetaKey permissionMetaKeyFromFastJson(FastJsonPermissionMetaKey fastJsonPermissionMetaKey);
+
+    WebInputPermissionMetaKey permissionMetaKeyToWebInput(PermissionMetaKey permissionMetaKey);
+
+    @InheritInverseConfiguration
+    PermissionMetaKey permissionMetaKeyFromWebInput(WebInputPermissionMetaKey webInputPermissionMetaKey);
 
     // -----------------------------------------------------------Rbac Entity-----------------------------------------------------------
     FastJsonPermission permissionToFastJson(Permission permission);
@@ -99,4 +113,14 @@ public interface BeanMapper {
 
     @InheritInverseConfiguration
     User userFromWebInput(WebInputUser webInputUser);
+
+    FastJsonPermissionMeta permissionMetaToFastJson(PermissionMeta permissionMeta);
+
+    @InheritInverseConfiguration
+    PermissionMeta permissionMetaFromFastJson(FastJsonPermissionMeta fastJsonPermissionMeta);
+
+    WebInputPermissionMeta permissionMetaToWebInput(PermissionMeta permissionMeta);
+
+    @InheritInverseConfiguration
+    PermissionMeta permissionMetaFromWebInput(WebInputPermissionMeta webInputPermissionMeta);
 }